chore: moved component
