+++ src/components/Photo/Photo.test.tsx
