+++ src/components/Photo/Photo.types.ts
