+++ src/components/Photo/index.ts
