+++ src/components/index.ts
@@ -1,2 +1,3 @@
 export * from './Header'
 export * from './List'
+export * from './Photo'
